fix cmd/clusterctl

+++ cmd/clusterctl/api/v1alpha3/metadata_type.go
@@ -36,10 +36,10 @@ type Metadata struct {
 // ReleaseSeries maps a provider release series (major/minor) with a API Version of Cluster API (contract).
 type ReleaseSeries struct {
 	// major version of the release series
-	Major uint `json:"major,omitempty"`
+	Major int32 `json:"major,omitempty"`
 
 	// minor version of the release series
-	Minor uint `json:"minor,omitempty"`
+	Minor int32 `json:"minor,omitempty"`
 
 	// contract defines the Cluster API contract supported by this series.
 	//
@@ -60,7 +60,7 @@ func init() {
 // GetReleaseSeriesForVersion returns the release series for a given version.
 func (m *Metadata) GetReleaseSeriesForVersion(version *version.Version) *ReleaseSeries {
 	for _, releaseSeries := range m.ReleaseSeries {
-		if version.Major() == releaseSeries.Major && version.Minor() == releaseSeries.Minor {
+		if version.Major() == uint(releaseSeries.Major) && version.Minor() == uint(releaseSeries.Minor) {
 			return &releaseSeries
 		}
 	}
